decimal capability 2

+++ src/transaction/AddIncome.js
@@ -152,7 +152,7 @@ return parts.join('.')
       console.log("updating");
       await updateDoc(doc(db, "Income", id.trim()), {
         // amount: parseInt((form.amount).split(' ').join('')),
-        amount: parseInt(form.amount),
+        amount: parseFloat(form.amount),
         date: form.date,
         desc: form.desc,
         incomeSource: form.income_source,
